Add ProGuard to Maven build
Note: Only for shrinking & optimizing.

This also fixes the Maven Shade Plugin build warning. That warning was the actual reason we'll considering to switch ProGuard anyway. There is a PR fixes that but there have been 15 days and it has not yet been merged. Switching to Gradle is not a goal at this point, but maybe in future.