Refactor of project structure
More refactorings are in the way.

Other changes made in the way with refactorings:
- Change version to 0.5 for next development iteration
- Added jetbrains-annotations as compile time dependency
- Added JUnit as a test scoped dependency
- Added maven-failsafe-plugin execution
- Now gets groupId and artifactId from project.properties,
allows for no-changes in source code if we change it on pom.
- Fixed a bug causing all compiler generated subclasses to be skipped.
This occurs because of a mistake in shouldSkip method.

- Cache ClassFileVersion objects to avoid Young Gen GCs. This will save
memory too. Before, if you process 20K classes, it would create 20K instances.
Now it will only create instances on different bytecode versions.

The current cache does not include minor version though, so preview classes, etc.
will not be cached.

- Replaced new ZipFile code in shouldSkip to use a single ZipFile instance.
This improves performance by a lot, for my testing, a whopping %100.
From 2.6s to 1.6s when processing a 130mb JAR file, with like 40k classes.

- Made some static variables lazy-init to optimize startup time.
- Refactored argument parsing again.

- Made ProgressTracker use a lower thread priority and use Thread#yield / Thread#onSpinWait
where possible to not block main task progression.

- Added verbosity and logging stuff to control message verbosity and logging.
Still not using a logging API though, this only internal API for our usage purposes.

- Added parallel processing: This will speed up processing time by utilizing all your processor cores.
This would result in a %100 speed up per cpu core, but my machine has only 2 cores so didn't able to get a
good improvement on my machine, but it will help when you have 3 or more cores.

- Added --fail-verbosity and --verbosity arguments.
- Other new arguments: async, buffered, help, parallel, threads, timing, track, verify